# clients/openapi.yaml
openapi: 3.1.0
info:
  title: Check Planner API
  description: This API is the Check Planner agents creator.
  version: 0.1.0

servers:
  - url: /check-planner/api/v1

paths:
  /:
    get:
      summary: Hello
      operationId: Hello
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloOutput"

  /generate:
    post:
      summary: Check Plan Generator
      operationId: Generate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_Check_Plan_Generator_generate_post"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResult"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

components:
  schemas:
    AgentResult:
      title: AgentResult
      type: object
      required:
        - rg_path
        - max_pages
        - max_rgs
        - output_file
      properties:
        rg_path:
          type: string
          title: Rg Path
        max_pages:
          type: integer
          title: Max Pages
        max_rgs:
          type: integer
          title: Max Rgs
        output_file:
          type: string
          title: Output File

    Body_Check_Plan_Generator_generate_post:
      title: Body_Check_Plan_Generator_generate_post
      type: object
      required:
        - reglements
      properties:
        reglements:
          title: Reglements
          type: array
          items:
            type: string
            format: binary

    HelloOutput:
      title: HelloOutput
      type: object
      properties:
        Message:
          type: string
          title: Message
          default: ""

    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"

    ValidationError:
      title: ValidationError
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type